Revert "float"
This reverts commit 22346e30395ef7fd5c88c392d4dd2001a3b7c84d.

+++ src/OnAllDesktopsButton.h
@@ -44,7 +44,7 @@ public:
         painter->setRenderHints(QPainter::Antialiasing, true);
         button->setPenWidth(painter, 1.2);
 
-        const qreal radius = 6;
+        const int radius = 6;
         painter->drawPolygon( QVector<QPointF> {
             QPointF(-radius, 0),
             QPointF(0, -radius),
